use the protobuf object directly for all operations on the directory list
also acknowledge the existence of an incomplete list on memcached
in that case, take values from the db and don't try to complete it.
assume that another thread already does it.

also don't continue with objects that are larger than 1MB, as we
could not store them on memcached.
this should also eliminate the possibility of filling local RAM
while building up the memcached object.
-> large directories will not be cached. :(