LCGDM-1246: specify http as protocol to do the dpm_put
dpm has a mechanism to _not_ use the automatic
internal file placement when a specific protocol is used.
http, https, and xrootd are hardcoded (all available
protocols supporting redirection).

By giving http as a possible protocol (and first one in
the list, so it will be chosen), we are allowed to write
on the location we want and the putdone will succeed.

It happens like this:
set protocol == http for dpm_put
dpm does _not_ set a pfn
we put the file wherever we want
dpm_putdone will look for an open put request
  for this file and create a replica entry