LCGDM-1246: provisionally add dmlite creation of put path
when dpm_put is invoked with protocol http, it
does not create a pfn (a physical path corresponding
to a server, filesystem, and filename in it)

We have to craft it ourselves. With only dpm it
should be possible to use anyhting we want; with
dmlite the dmlite token must match.

So I use the implementation that I used earlier
that creates a path and also uses this to craft
the dmlite token.

This is just a proof-of-concept, as it needs a
dmlite catalog to function, which we don't want.