LCGDM-1246: make dpm_put with the physical file name
The mechanism used in the earlier commits needs a
change in how the dpm_put is involved.
The putdone later matches the sfn provided to a replica
and completes if the replica has a corresponding
putfile request.

This means we have to do the put with the physical
file name as sfn so that the replica filename and
filename (sfn) in the put request match.

This allows to use an external system to put a file
and just inform DPM later on to import that file
into its catalog

it's:
1. dpm_put(phys. filename)
2. create replica entry(kAvailable)
3. invoke dpm_putdone(phys. filename)