LCGDM-1246: use teh 'real' sfn in the redir request
In putdone, the replica entry is searched for by sfn,
apparently in a complicated manner. In the end, dpm needs
an sfn entry that also exists a valid file in the namespace.

Then the replica sfn, which is the phys. filename, is used to
find the put request (which will work), and everything should
succeed nicely.