LCGDM-1246: implement the new idea
first: it will not compile without the new dpm_putx
(or similar) function.

The idea is the following:
Dmlite either
1. selects a specific filesystem from the queryString or
2. selects a random fs if a pool has been specified in the
  queryString or
3. lets DPM decide on both

In the first two cases, it sends the selects filesystem
as queryString param to the surl to dpm_putx, a modified
dpm_put function. This will interpret the parameter of form
<server>:<fs> and return a complete transfer URL (to the
phys. file) to dmlite, taking the filesystem constraints
into account.

Dmlite already makes sure that the pool and filesystem match!
The Catalog creates a PoolDriver for a specific pool, then the
PoolDriver creates a PoolHandler and this one calls the
dpm_getpoolfs function with the pool as an argument, so
it only returns filesystems within that pool.