LCGDM-1246: send the replicate flag to dpm_put
If the queryString parameter replicate is set, then
we give the replicate flag to dpm_put and dpm_putx.
This way, dpm does not create a new namespace entry,
but just a new replica.

Combined with a later COPY with the returned URL
as destination, this allows us to implement the
dpm-replicate command
1. with a dpm-chosen destination
2. with a user-chosen pool and/or filesystem