LCGDM-1257: add token w/o user credentials
If you specify the replicate flag in a put, dmlite creates a token
which is not related to your user id so you can use it later for
the copy from a not-yet-known disk server.

The receiving end then tries if a user-Id-dependent token can be
verified and also for the non-dependent.