LCGDM-1311: provide inode function updateextendedattributes in the adapter
Add the inode interface to the Adapter loaded with the NsAdapter plugin.
It only provides two functions: getImplId and 
updateExtendedAttributes, all others give a notimplemented
exception.

The function updateExtendedAttributes takes an inode
and calls the dpns function Cns_getpath to get the path
and then calls the catalog update function with it.

This first version depends on the env var DPNS_HOST.