fix the creation of a memcached conn in certain conditions
Sometimes, depending on the config file contents, the
memcached conenctison were not able to interact with the
cache servers.

This is probably due to a missing string terminating \0
in the value that holds the memcache server name.
The memcpy, which creates this value, has been replaced
with a strncpy(.., .., n=src_length+1) to make sure that
the last character is \0.
strncpy automatically pads with \0 if n > src_length.
dst has the size src_length+1 as well.