fix creation of timestamped result filename
Relative paths to the result file were altered, so that
the timestamped copy was likely not to be created.

We have the path of the csv file, in a specific case like this:
./results/blablabla.csv

This was copied and altered to create:
./results/blablabla_<timestamp>.csv

First '.csv' was removed with .strip('.csv'), which also
removed the . at the beginning.

Since we know that the file ends with .csv, I changed this
to just remove the last 4 characters:
orig_file_name[:-4]