LCGDM-1277: redo the argus variable management
Until now, argus was built even when it was disabled in the
spec file. This failed then due to missing dependencies.

So I remodeled the mock config and spec stuff after this:
http://rpm5.org/docs/api/conditionalbuilds.html

The mock config creates the _with_argus macro, which is
picked up in the spec and builds▫the extra argus rpm.

Additionally, the _without_argus macro is passed to
configure, so argus is not built, when we don't have
the dependencies.

Deviation from the link: our ./configure does not accept
--with-argus, so this is not passed.