LCGDM-1337: use the connection pool
So far the memcached connection have been created on
every instantiantion of Catalog.
Log messages show that, on a loaded machine, this happens
roughly every 5 seconds, so many connection are uselessly
created and destroyed.

This commit goes back to using the dmlite ConnectionPool,
which now does not have a strict limit on the number of
connections, but lets it grow dynamically
(with another limit of connections at 2*n).

The default number of connections has been changed to 100.

The missing connection reuse might be a cause for the
unexpectedly low performance of the memcache plugin.

1337 commit, lol ;)