LCGDM-1337: bump the memc connection number to 500
The current performance tests with 100 - 500 clients
show a drop (half the performance,~450 to 200) going
from 200 clients to 300.

This might be due to the fact that the PoolContainer
does not allow more than 200 open connections.

Since every Catalog instance needs 2 connections, it
would mean that for a perfsuite test with n clients,
n/2 connections are used in parallel.