fix syslog in memcachefunctioncounter
It erroneously printed on every function execution, even when there
was no data available. It then put a blank line.

This happened when trying to put the syslog statement
outside of the mutexed part.
I refactored so that the mutex lives in an additional
block ({ }).
Hope it doesn't get optimized away. If so, the
syslog statement is also protected by the mutex ...