LCGDM-1338: correct creation of paths in memcache readdirx
So far memcache expected a directory path ending with a slash,
which lcgdmdav provides, but dpm-dsi does not, so
gridftp listings failed.

This introduces a new method, concatPath, to join the dir
with its entries, that check whether the dir entry has a slash
and adds it if appropriate.

Another solution could be to use Url::normalizePath and add
a slash in any case, but this seems more expensive.
concatPath only checks for the '/', then return a newly
created string.
normalizePath splits the string in a vector and concats the
parts, getting rid of double slashes.