guarantee that MemcacheConnectionFactory function never throw
Alter the create method to replace all exceptions with syslog writes.
Failed connection to memcache should not be fatal, but be noticed by the admin.

Throwing exceptions apparently has influences on the delegated class as well
and can lead to resource leaks. Case:
Connections of the underlying mysql class were not destroyed when the
MemcacheConnectionFactory::create threw an exception that was uncaught.

Added no throw guarantees to create(), destroy(), and isValid().