LCGDM-1352: use a central counter to divide the work between threads
The threads used to have a statically allocated number of files to
process. Now they get the next file to use from a central counter
(representing the file numbering schema) and finish when there is no
more work to do.

This should give more accurate results, as we don't wait anymore
for the slowest thread to finish.