LCGDM-1355: only save the rfns in the replica list in cache
Change the replica list to behave like the dir list in the cache.
The list only ohlds the keys to its entries, not the entries itself.

This harmonizes the way replicas are cached: before a replica
would exist in the replicas list and might also exist as a single
one identified by rfn.
Now the list only points to the rfn entry.

This is good, because it allows large lists (memcache limit 1MByte).
It's bad, because it slows down getting the replica list.
One get, versus 1 list get + n entry gets.

This should be avoided with a multiget, but I won't implement it
before it shows to be a real performance problem.