LCGDM-1355: fall back to db, if all replica list entries fail
change the behaviour of getReplicas.
If all replicas of a file have been deleted and thus their cached
representations, too, fall back to mysql.
This will catch new replicas in the case that a file has been overwritten.

However, that also means that a call to the db is triggered,
when all replicas have been deleted. the db call then will fail.
This should not make too much problems, because I except that
when all replicas are deleted, the namespace entry would be
deleted, too. Then, the replica list can also be removed by
the memcache plugin.