move the davfile construction one level up
move the creation of the davfile object so that it
is reused when doing multiple requests to a URL.
This might benefit the results when -q > 1.

A quick analysis with mutrace has shown that instantiating
the NEON session in davix creates contention on a mutex.
Since the mutex was used more than 40k times on a test
with 10k files and -q 2, I suspect that it's used in the
davfile creation process.

This is just a test, and it might well be that there is no
impact on performance at all.