add a script to check results
the (python) script checks for strange outcomes of the test runs and
exits with sys.exit(1) to fail the job.

It checks for very low throughput (<= 10 req/sec) or if all files
of all runs of a test have failed.

This should alert us by failing the job that something is wrong
with either the test or with the endpoints.