send the userident message on the first API call used
the userident msg cannot be called on setSecurityContext because
it is also called on the instantiation of the stackinstance with
'root' context.
We then get two messages, one for 'root' and one for the correct
user when the credentials are set.

The new method is to potentially send the userident msg on all
API function calls, but the first call sets a marker in the
stackinstance so that further invokations do not send another
msg.

see sendUserIdentOrNOP() around the ProfilerCatalog