change the counter that points to the next free entry in the buffers
it used to point to 'slots (used) + 1' which is wrong when we count
from 0. the first entry with size 8 slots occupies slots 0-7, so
the second entry should start at slot 8 (not 9), and so on ..

I think this led to the failing parse of the fstream, and, if true,
would also mean that the gled collector does not report failed
parsing of the redir stream.