move the XrdMon related function in Profiler* to ProfilerXrdMon
the different parts of dmlite (Catalog, Pool*, and IO) need
the same XrdMon-related functions: sendUserIdent, report***, and so on.
They work on the stackinstance to save state, and thus cannot
be included in the static XrdMonitor class.

They are now put in ProfilerXrdMon, from which all interfaces
inherit. Because we need the stack, it also moves up into the new class.