keep the dictid consistent for a user
my understanding of the dictid idea might be false up until now,
so we'll try this new things.

A dictid mapping saves dictids, which allows us to find back
the dictid of a user even between stackinstances, e.g. when
going from the catalog to the poolmanager, or io.

so far,
1. the dictid stays until eternity
  depending on how it really works, we need to delete them.
2. only the username (DN) is taken into account; in the end
  this should also include the client host, maybe more info.