allow up to two collectors
all packets will be sent to both, sequentially.
this is semantically different from xrootd, as they allow different options
for different collectors
and it's performancially (...) different as it will slow down sending messages a lot.

let's see in performance testing, this now is the simplest possible implmentation,
motivated by sending packets to an official gled collector and also to ours,
so we can continue looking at the logs.