fix assigning multiple collectors
the dmlite configure method goes rampage. it is called many, many times
(12 in the test) for the same parameters, i.e. 'Collector'.
It you specify one Collector, it gets added 12 times, leading to problematic
multiple packets.

Change the container to store the collector names to a set, so we
avoid duplicate entries in any case.