fix sending Disconnect messages correctly
1. rename Disc to DiscAndFlush to indicate that it flushed the stream
2. only do something if a dictid is available in the stack
  -> new function: ...OrNOP

The function is called on the destruction of the IODriver, which
indicates that the client connection has terminated.
Apparently the dstructor is also sometimes called without any
connection being open, so bogus messages (with bogus IDs)
have been sent.
Now we only send and flush if a dictid is on the stack,
do nothing otherwise.