LCGDM-1483: add a more efficient extendedStat function in memcache plugin
add a function that can resolve paths more quickly, without using a vector of string,
that for the memcache plugin has to be constructed into a string on every level.
It works for paths without '.', '..' or symlinks, for clarity, and falls back to the
slow version if any of this is encountered. Because these cases should not happen
often, there should be a general speedup.
On the raspberry, it gave 20% speedup for the stat test.