LCGDM-1464: fix expiration of old values
do not update an item's time on reuse, so that even old, but often-used
values can be expired. the LRU value of each item is now only determined
through its place in the list, the time is always the time of insertion.
thus the expire algorithm has to check every single list item if it
expired. now it's clearly O(n) with n comparisons.