LCGDM-1504: stop using the stackinstance in the IOHandler
ProfilerXrdMon, the parent class of ProfilerCatalog, -PoolManager,
and -IO, now works with both a StackInstance set or with the two
variables secCtx_ and protocol_.
The ProfilerIO sets those two, the StackInstance is never set. It
is set to 0x00 as default to allow testing its presence.

The dictid and fileid now staty strictly within the Catalog,
PoolManager, or Profiler. This should be fine with the current
messages used, and any communication probably wouldn't work
anyway, because it's difficult to rely on the stack's presence.

The function that should catch missed Close() messages in the
IODriver is removed, it wouldn't work anymore. This also removes
the need to let it inherit from ProfilerXrdMon.
This function was likely never used.

This first implementation does not feel completely right, it has
this invisible contract between ProfilerIO and ProfilerXrdMon to
set the required variables -- this can easily break.