reset counters after file close; small refactor
with connection keepalive the IOHandler will stay alive
for multiple files. we have to reset the monitoring counters
in between. resetCounters() does that and is called on every
close.

simplify the rmDictId function in ProfilerXrdMon.

stop flushing the file buffer on every file close, running at
disconnection should be enough.