LCGDM-1537: support the protocol field
The user can be described by: prot/user.pid:sid@host

prot is the communication protocol and will be set to whatever
the frontend gives us -- if nothing, to an empty string.

the protocol as in info: p=ap is either 'none', if the secCtx
cannot provide us with a username, of 'gsi' if it can.